refactor : piece에서 posotion 삭제
