+++ src/main/java/piece/Bishop.java
@@ -1,21 +1,17 @@
 package piece;
 
-import java.util.Objects;
 import model.Camp;
 import point.Position;
 
 public class Bishop extends Piece {
 
-    private Position position;
-
     public Bishop(final Camp camp) {
         super(camp);
     }
 
-
     @Override
     void move(Position targetPosition) {
-        this.position = targetPosition;
+
     }
 
     @Override
@@ -24,22 +20,5 @@ public class Bishop extends Piece {
             return "b";
         }
         return "B";
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Bishop bishop = (Bishop) o;
-        return camp == bishop.camp && Objects.equals(position, bishop.position);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(camp, position);
     }
 }
