+++ src/main/java/piece/King.java
@@ -1,13 +1,9 @@
 package piece;
 
-import java.util.Objects;
 import model.Camp;
 import point.Position;
 
 public class King extends Piece {
-
-    private Position position;
-
 
     public King(final Camp camp) {
         super(camp);
@@ -16,7 +12,7 @@ public class King extends Piece {
 
     @Override
     void move(Position targetPosition) {
-        this.position = targetPosition;
+
     }
 
     @Override
@@ -25,22 +21,5 @@ public class King extends Piece {
             return "k";
         }
         return "K";
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        King king = (King) o;
-        return camp == king.camp && Objects.equals(position, king.position);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(camp, position);
     }
 }
