+++ src/main/java/piece/Knight.java
@@ -1,12 +1,9 @@
 package piece;
 
-import java.util.Objects;
 import model.Camp;
 import point.Position;
 
 public class Knight extends Piece {
-
-    private Position position;
 
     public Knight(final Camp camp) {
         super(camp);
@@ -14,7 +11,6 @@ public class Knight extends Piece {
 
     @Override
     void move(Position targetPosition) {
-        this.position = targetPosition;
     }
 
     @Override
@@ -23,22 +19,5 @@ public class Knight extends Piece {
             return "n";
         }
         return "N";
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Knight knight = (Knight) o;
-        return camp == knight.camp && Objects.equals(position, knight.position);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(camp, position);
     }
 }
