+++ src/main/java/piece/Piece.java
@@ -1,10 +1,10 @@
 package piece;
 
+import java.util.Objects;
 import model.Camp;
 import point.Position;
 
 public abstract class Piece {
-
 
     protected final Camp camp;
 
@@ -16,5 +16,22 @@ public abstract class Piece {
 
     public Camp getCamp() {
         return camp;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Piece piece = (Piece) o;
+        return camp == piece.camp;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(camp);
     }
 }
