+++ src/main/java/piece/Queen.java
@@ -1,12 +1,9 @@
 package piece;
 
-import java.util.Objects;
 import model.Camp;
 import point.Position;
 
 public class Queen extends Piece {
-
-    private Position position;
 
     public Queen(final Camp camp) {
         super(camp);
@@ -14,7 +11,6 @@ public class Queen extends Piece {
 
     @Override
     void move(Position targetPosition) {
-        this.position = targetPosition;
     }
 
     @Override
@@ -23,22 +19,5 @@ public class Queen extends Piece {
             return "q";
         }
         return "Q";
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Queen queen = (Queen) o;
-        return camp == queen.camp && Objects.equals(position, queen.position);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(camp, position);
     }
 }
